infrastructure/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.2
idna==3.4
iniconfig==2.0.0
lark==1.1.5
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
python-hcl2==4.3.2
requests==2.31.0
tomli==2.0.1
urllib3==2.0.3
